refact: moving files/folders to better places
