+++ test/vaultscan/e2e/test_cli_find_command.py
