+++ test/vaultscan/e2e/test_cli_vault_command.py
