+++ test/vaultscan/e2e/test_cli_version_command.py
