+++ test/vaultscan/unit/vaultscan/repositories/vault/test_factory.py
