+++ test/vaultscan/unit/vaultscan/util/test_os.py
