+++ vaultscan/repositories/vault/base.py
@@ -13,7 +13,6 @@ class VaultStatus(Enum):
 class BaseVaultConfig:
     alias: str
     type: str = field(init = False)  # Each subclass should set this
-    # FIX IT: Change it to VaultStatus
     status: str
 
     def __post_init__(self):
@@ -28,6 +27,7 @@ class BaseVaultConfig:
 
 
 class VaultRepository(ABC):
+    ''' Base class to repository of vaults '''
     @abstractmethod
     def initialize(self): pass
